+++ src/test/java/de/chrgroth/jsonstore/store/JsonSingletonStoreTest.java
@@ -26,10 +26,10 @@ public class JsonSingletonStoreTest {
     public void init() {
         tempDir = Files.createTempDir();
         FlexjsonHelper flexjsonHelper = FlexjsonHelper.builder().dateTimePattern(DATE_TIME_PATTERN).build();
-        persistentStore = new JsonSingletonStore<>(String.class, null, flexjsonHelper, tempDir, StandardCharsets.UTF_8, true, true);
-        persistentStoreCopy = new JsonSingletonStore<>(String.class, null, flexjsonHelper, tempDir, StandardCharsets.UTF_8, true, true);
-        transientStore = new JsonSingletonStore<>(String.class, null, flexjsonHelper, null, null, false, false);
-        transientStoreCopy = new JsonSingletonStore<>(String.class, null, flexjsonHelper, null, null, false, false);
+        persistentStore = new JsonSingletonStore<>(String.class, null, flexjsonHelper, tempDir, StandardCharsets.UTF_8, true, true, false);
+        persistentStoreCopy = new JsonSingletonStore<>(String.class, null, flexjsonHelper, tempDir, StandardCharsets.UTF_8, true, true, false);
+        transientStore = new JsonSingletonStore<>(String.class, null, flexjsonHelper, null, null, false, false, false);
+        transientStoreCopy = new JsonSingletonStore<>(String.class, null, flexjsonHelper, null, null, false, false, false);
         testDataOne = "test data";
     }