+++ RELEASE-NOTES.md
@@ -3,6 +3,9 @@ Release Notes
 
 0.8.0 (in progress)
 -------------------
+- fixed bug if a custom handler for java.util.Date or java.time.LocalDateTime is registered
+- added methods to add handlers for specific store only
+- added deep serialization mode to json stores, see flexjson documentation for detailed information
 - ...
 
 0.7.1